+++ bootique-undertow-example/src/main/java/io/bootique/undertow/Application.java
@@ -4,9 +4,6 @@ import com.google.inject.Binder;
 import com.google.inject.Module;
 import io.bootique.Bootique;
 
-import static io.bootique.undertow.UndertowModule.contributeControllers;
-import static io.bootique.undertow.UndertowModule.contributeWrappers;
-
 public class Application implements Module {
     public static void main(String[] args) {
         Bootique.app(args)
@@ -17,12 +14,9 @@ public class Application implements Module {
 
     @Override
     public void configure(Binder binder) {
-        contributeControllers(binder)
-                .addBinding()
-                .to(EchoRestController.class);
 
-        contributeWrappers(binder)
-                .addBinding()
-                .to(ApplicationExceptionHandler.class);
+        UndertowModule.extend(binder)
+                .addController(EchoRestController.class)
+                .addWrapper(ApplicationExceptionHandler.class);
     }
 }