+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 0.2
 
 * #1 Upgrade to Bootique 0.22
+* #2 Replace "contribute*" methods with more modern "extend" API
 
 ## 0.1